Move common logic to common classes
