+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfileContract.kt
@@ -2,30 +2,22 @@ package com.bounswe2020group3.paperlayer.profile
 
 import com.bounswe2020group3.paperlayer.base.BasePresenter
 import com.bounswe2020group3.paperlayer.base.BaseView
+import com.bounswe2020group3.paperlayer.profile.common.ProfileCommonContract
 import com.bounswe2020group3.paperlayer.profile.data.Profile
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 interface ProfileContract {
-    interface Presenter: BasePresenter {
-        fun fetchProfile(profileId: Int)
+    interface Presenter: BasePresenter, ProfileCommonContract.Presenter {
     }
 
-    interface View: BaseView<Presenter> {
-        fun showLoading()
-        fun hideLoading()
-        fun showInfoToast(message: String = "Info")
-        fun showErrorToast(message: String = "Error")
-        fun updateProfileUI(profile: Profile?)
+    interface View: BaseView<Presenter>, ProfileCommonContract.View {
     }
 
-    interface Model {
-        fun getProfile(profileId: Int): Single<Profile>
+    interface Model: ProfileCommonContract.Model {
     }
 
-    interface ProfileService {
-        @GET("/api/profiles/{profileId}/")
-        fun getProfile(@Path("profileId") profileId: Int): Single<Profile>
+    interface Service: ProfileCommonContract.Service {
     }
 }
