+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfilePresenter.kt
@@ -1,35 +1,14 @@
 package com.bounswe2020group3.paperlayer.profile
 
-import android.util.Log
-import io.reactivex.disposables.CompositeDisposable
+import com.bounswe2020group3.paperlayer.profile.common.ProfileCommonPresenter
 
-class ProfilePresenter(view: ProfileContract.View) : ProfileContract.Presenter {
+open class ProfilePresenter(view: ProfileContract.View) : ProfileCommonPresenter(view), ProfileContract.Presenter {
 
     private var view: ProfileContract.View? = view
-    private var model: ProfileContract.Model = ProfileModel()
-
-    private var disposable = CompositeDisposable()
 
     override fun onDestroy() {
+        super.onDestroy()
         this.view = null
-        disposable.clear()
-    }
-
-    override fun fetchProfile(profileId: Int) {
-        view?.showLoading()
-        view?.showInfoToast("Fetching the profile...")
-        val getProfileObservable = model.getProfile(profileId).subscribe(
-                { profile ->
-                    view?.updateProfileUI(profile)
-                    view?.hideLoading()
-                    view?.showInfoToast("Profile is fetched.")
-                },
-                { error ->
-                    view?.hideLoading()
-                    view?.showErrorToast("Error while fetching profile")
-                }
-        )
-        disposable.add(getProfileObservable)
     }
 
 }
